Add an authentication blueprint called auth.
Import werkzeug.security from flask and set up two functions, one to generate a password and one to check the password against the generated hash. Also, setting up a test for the password and registering the authentication blueprint to the create_app function that initializes the entire app in app/__init__.py file.